+++ SPO2/UpperLevel/Util/Util.h
@@ -26,7 +26,7 @@ uint16_t POWC(uint32_t a,uint8_t b);
 /*¬озврат строкового аналога числа (от 0 до 9)*/
 uint8_t intToChar(uint8_t num);
 
-uint8_t* intToStr (int32_t num);
+uint8_t* intToStr (int64_t num);
 
 uint32_t mod(int32_t a);
